requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
yfinance>=0.2.28
ta-lib>=0.4.27
